% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papersize.R
\name{aseries}
\alias{aseries}
\alias{bseries}
\alias{papersize}
\title{Determine the width and height of the ISO216 A and B series of paper}
\usage{
aseries(n, floor = TRUE)

bseries(n)
}
\arguments{
\item{n}{is an integer from 0 to 10}

\item{floor}{returns an integer by taking the floor of the result}
}
\value{
a vector containing the width and height
}
\description{
Determine the width and height of the ISO216 A and B series of paper
}
\section{Functions}{
\itemize{
\item \code{aseries}: A paper size

\item \code{bseries}: B paper sizes
}}

\examples{
aseries(4) # Return A4 size paper
bseries(5) # Return B5 size paper
}
