% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dawn_dusk.R
\name{calculate_dawn_dusk}
\alias{calculate_dawn_dusk}
\title{Determine dawn and dusk for a specified GPS coordinate.}
\usage{
calculate_dawn_dusk(datetime, gpscoord)
}
\arguments{
\item{datetime}{the date-time to determine when dawn and dusk occurs.}

\item{gpscoord}{the gps coordinate to determine when dawn and dusk occurs. It should be provided as a 1x2 matrix, where the first value is the latitude and the second value is the longitude.}
}
\value{
A lubridate interval.
}
\description{
Determine dawn and dusk for a specified GPS coordinate.
}
