% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rate.R
\name{calculate_rate}
\alias{calculate_rate}
\title{Calculate the time rate-of-change in dissolved oxygen concentrations.
This function will fit a GAM assuming a scaled t distribution.}
\usage{
calculate_rate(data, response, k = 24, bs = "cr")
}
\arguments{
\item{data}{is a tibble containing one day of oxygen data.}

\item{response}{is the name of the variable containing the oxygen data.}

\item{k}{= 24 is the number of basis functions used to fit the GAM.}

\item{bs}{= "cr" the type of basis function to use in the GAM.}
}
\value{
A tibble of the data passed to the function is returned, including the fitted values and the derivatives.
}
\description{
Calculate the time rate-of-change in dissolved oxygen concentrations.
This function will fit a GAM assuming a scaled t distribution.
}
