% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_mgl.R
\name{fix_mgl}
\alias{fix_mgl}
\title{Adjust dissolved oxygen concentrations
Use temperature and salinity to adjust the oxygen concentration.}
\usage{
fix_mgl(mgl, temperature, salinity = 32)
}
\arguments{
\item{mgl}{is the dissolved oxygen concentration in mg / L}

\item{temperature}{is the water temperature in celcius}

\item{salinity}{= 32 is the salinity in PSU}
}
\description{
Adjust dissolved oxygen concentrations
Use temperature and salinity to adjust the oxygen concentration.
}
