% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_mgl.R
\name{fix_mgl2}
\alias{fix_mgl2}
\title{Adjust dissolved oxygen concentrations
Use barometric pressure, temperature, and salinity to adjust the oxygen concentration.}
\usage{
fix_mgl2(mgl, temperature, bar, salinity = 32)
}
\arguments{
\item{mgl}{is the dissolved oxygen concentration in mg / L}

\item{temperature}{is the water temperature in celcius}

\item{salinity}{= 32 is the salinity in PSU}

\item{pressure}{is the barometric pressure in HPa}
}
\description{
Adjust dissolved oxygen concentrations
Use barometric pressure, temperature, and salinity to adjust the oxygen concentration.
}
