% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_transfer.R
\name{masstransfer}
\alias{masstransfer}
\title{Calculate the mass transfer rate of oxygen at the air-sea interface.
This version ignores barometric pressure.}
\usage{
masstransfer(windspeed, temperature, salinity, oxygen, height = 1)
}
\arguments{
\item{windspeed}{is the windspeed in m / s}

\item{temperature}{is the water surface temperature in Celsius}

\item{salinity}{is the salinity in PSU}

\item{oxygen}{is the dissolved oxygen concentration at the surface in mg / L}

\item{height}{= 1 is the height of the wind sensor in m}
}
\value{
a vector of fluxes with units g / m2 / hr. Positive values indicate flux out of the water surface and negative values indicate flux into the water surface.
}
\description{
See Crusius and Wanninkhof 2003 Limnology and Oceanography 48 for details
}
