% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_awh.R
\name{read_awh}
\alias{read_awh}
\title{Read ALEC AWH data
This function can read the AWH data files.
Note that `options(digits.secs = 4)` might needed to be set to view the subseconds.}
\usage{
read_awh(filename, summarisedata = T, ...)
}
\arguments{
\item{filename}{the name of the file which the data are to be read from. It must be a CSV file}

\item{summarisedata}{has a default value of FALSE. If set to TRUE, the burst samples will be averaged.}

\item{...}{further arguments passed to read_csv}
}
\value{
A tibble containing the data.

\strong{Returned variables}
\describe{
   \item{datetime}{Date-time}
   \item{mpa}{Pressure (MPa)}
   \item{depth}{Depth (m)}
}
}
\description{
Read ALEC AWH data
This function can read the AWH data files.
Note that `options(digits.secs = 4)` might needed to be set to view the subseconds.
}
\examples{

}
