% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_castaway.R
\name{read_castaway}
\alias{read_castaway}
\title{Read Castaway CTD data}
\usage{
read_castaway(filename, skip = 29, ...)
}
\arguments{
\item{filename}{the name of the csv file.}

\item{skip}{the number of lines to skip. Default is 29.}

\item{...}{further arguments passed to read_csv.}
}
\value{
A tibble containing the data.

\strong{Returned variables}
\describe{
  \item{lat}{GPS latitude}
  \item{lon}{GPS longitude}
  \item{datetime}{Date and time of the cast}
  \item{depth}{Water depth (m)}
  \item{temperature}{Water temperature (\eqn{^\circ}{degrees}C)}
  \item{salinity}{Salinity (PSU)}
  \item{density}{Water density (kg / m^3)}
}
}
\description{
This function reads the CTD from the Castaway CTD.
The data that is read is the mean start and ending GPS coordinates,
the time of the cast, water depth, water temperature, salinity, and water density.
}
