% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_odyssey.R
\name{read_odyssey}
\alias{read_odyssey}
\title{Read Odyssey data
This function reads the Odyssey PAR datalogger data files.}
\usage{
read_odyssey(filename, ...)
}
\arguments{
\item{filename}{the name of the file which the data are to be read from. It can be a CSV file or an XLSX file.}

\item{...}{further arguments passed to read_csv or read_xlsx}
}
\value{
A tibble containing the data.

\strong{Returned variables}
\describe{
\item{`datetime`}{Date-time}
\item{`ppfd`}{Uncalibrated photosynthetically active radiation (-)}
}
}
\description{
When the PAR value is below 0, the value is set to NA.
**At this stage, the data is not calibrated!**
}
\examples{
\dontrun{
fnames = dir("~/Lab_Data/kawatea/", full.names=TRUE, recursive=TRUE)
df = tibble(fnames = fnames)#'
df \%>\% filter(str_detect(fnames, "Light")) \%>\% slice(1) \%>\% pull(fnames) \%>\% read_odyssey()
}

}
