% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save ggplot to pdf and png file.}
\usage{
save_plot(
  pdfname,
  plot = last_plot(),
  width = 100,
  height = 100,
  units = "mm",
  density = 300
)
}
\arguments{
\item{pdfname}{the pdf filename}

\item{plot}{the ggplot object or the most recent figure in the plot panel.}

\item{width}{the width of the saved plot in mm. The default is 100.}

\item{height}{the height of the saved plot in mm. The default is 100.}

\item{units}{the units of the plot dimensions. The default is "mm".}

\item{density}{the density of the png file in pixels. The default is 300.}
}
\description{
Saving a ggplot into a pdf then converting the pdf into a
png file. The filenames will automatically be saved with
a pdf and png extension.
}
\examples{
\dontrun{
save_plot("filename.pdf", width = 100, height = 100, units = "mm")
}

}
