% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_functions.R
\name{shift_months_general}
\alias{shift_months_general}
\title{Shift months
Shift the months by subtracting the shift degree
A positive shift moves months forward, while a negative shift moves them backward}
\usage{
shift_months_general(original_month, shift_month)
}
\arguments{
\item{original_month}{is the original month of the data}

\item{shift_month}{is the degree of shifting}
}
\description{
Shift months
Shift the months by subtracting the shift degree
A positive shift moves months forward, while a negative shift moves them backward
}
